Change logfile test for tcsh to strip() output string
There is a difference between tcsh in CentOS-5 vs. 6 that introduces
a difference in the way pexpect gets tcsh output.  This ends up generating
an extra \r\r\n in the CentOS-6 case.  No way around this was found in
the allotted time, so it was decided to accept that the logfile output
in CentOS-6 will just have an extra line at the top.  The test was changed
to ignore that difference.